LayoutTests/fast/regex/dom/cross-frame-callable.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
<script>
function doTest(childRegExp)
{
    re = childRegExp;
    shouldThrow("re('a')");
    shouldBe("re.exec('a')", "['a']");
}

var iframe = document.createElement('iframe');
document.body.appendChild(iframe);
iframe.contentDocument.write('<script>top.doTest(/a/)</scr' + 'ipt>');
iframe.contentDocument.close();
document.write('DONE');
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
