LayoutTests/fast/regex/dom/cross-frame-callable-expected.txt

PASS re('a') threw exception TypeError: RegExp is not a function (evaluating 're('a')').
PASS re.exec('a') is ['a']
PASS successfullyParsed is true

TEST COMPLETE
DONE
